updated code

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Post;
+use App\Models\User;
 use App\Models\Category;
 use Validator;
 
@@ -14,29 +15,97 @@ class PostController extends Controller {
 	        return view('auth.login');
 	    }       
     } 
+	
+	
+	public function home(Request $request)
+    {
+		$terms = explode(",", request('q'));
+		$q = request('q');
+		$posts = array();
+		if(isset($q) && !empty($q)){			
+			$posts = Post::query()	
+			->where('status', 1)	
+			->where(function ($query) use ($terms) {
+				foreach ($terms as $term) {
+					$query->orWhere('tags', 'like', '%' . $term . '%');
+				}
+			})
+			->paginate(5);
+			$posts->appends (array ('q' => $q));	
+		}else{
+			//$posts = Post::where('status', 1)->paginate(1);
+        } 
+		//$posts = Post::latest()->paginate(5); 
+		return view('home',compact('posts'))
+            ->with('i', (request()->input('page', 1) - 1) * 5);		
+		
+		
+    }
 
     public function index()
     {	
-		$user_id = Auth::user()->id;	
-		$posts = Post::where('createdBy', $user_id)->paginate(5);
-        return view('frontend.posts.index',compact('posts'))
+	$user = Auth::user();		
+        if(Auth::check()){		
+		$user_id = Auth::user()->id;			
+		$user = User::find($user_id);       
+		$posts = Post::where('user_id', $user_id)->paginate(5);
+        return view('post',compact('posts','user'))
             ->with('i', (request()->input('page', 1) - 1) * 5);
+		}    
+		return redirect("login")->withSuccess('You are not allowed to access');	
     }
    
-	public function add(Post $post)
+	public function add()
     {
-		 return view('frontend.posts.edit', [
-            'post' => $post,
-            'category' => Category::all()->pluck('name', 'id'),          
-        ]);
+		$data = Post::select("title","tags")
+		->where('status', 1)	
+		->get()		
+		->toArray();	
+		$newArray = array();
+		foreach($data as $datas){		
+			$newTags = explode(',', $datas['tags']);	
+			$newArray['val'][] = trim($datas['title']);		
+			foreach($newTags as $newTag){		
+				$newArray['val'][] = trim($newTag);	
+			}	
+		}		
+		$newArrayUnique = array_unique($newArray['val']);
+		$page = 'post';
+		$user = Auth::user();		
+        if(Auth::check()){		
+			$user = User::find($user->id);       
+			return view('post_add')->with(compact('user','newArrayUnique','page'));
+		}    
+		return redirect("login")->withSuccess('You are not allowed to access');
+		
+		 
     }	
 	
-	public function edit($id = null) {
-        $post = Post::find($id);
-         return view('frontend.posts.edit', [
-            'post' => $post,
-            'category' => Category::all()->pluck('name', 'id'),          
-        ]);		
+	public function edit($slug = null) {
+		$data = Post::select("title","tags")
+		->where('status', 1)	
+		->get()		
+		->toArray();	
+		$newArray = array();
+		foreach($data as $datas){		
+			$newTags = explode(',', $datas['tags']);	
+			$newArray['val'][] = trim($datas['title']);		
+			foreach($newTags as $newTag){		
+				$newArray['val'][] = trim($newTag);	
+			}	
+		}		
+		$newArrayUnique = array_unique($newArray['val']);
+		$page = 'post';
+		$user = Auth::user();		
+        if(Auth::check()){	
+			 $post = Post::where('user_id', $user->id)
+            ->Where('slug', $slug)
+            ->firstOrFail();				
+			$user = User::find($user->id);   
+			return view('post_edit')->with(compact('user','post','newArrayUnique','page'));
+		}    
+		return redirect("login")->withSuccess('You are not allowed to access');
+			
     }	
 	
 	public function update(Request $request, Post $post) {
@@ -44,26 +113,40 @@ class PostController extends Controller {
             'title' => 'required',
             'content' => 'required',
         ]); 
-		
-		if($request->input('id') != null){	
-			//unset($request->input('_token'));
-			$id = $request->input('id');
-		
+		unset($request['files']);
+		if($request->input('id') != null){				 
+			$id = $request->input('id');		
 			$post = Post::where('id',$id);
 			$data = $request->except('_token');
 			$post->update($data);			
-			 return redirect()->route('posts')
+			return redirect()->route('post')
                         ->with('success','Post updated successfully.');
 		}else{
-			$user_id = Auth::user()->id;	
-			$request['createdBy']	= $user_id; 
-			$request->except('name');
-			Post::create($request->all());
-			 return redirect()->route('posts')
+			$user_id = Auth::user()->id;
+			$data = $request->except('_token');
+			$data['user_id'] = $user_id; 
+			Post::create($data);
+			 return redirect()->route('profile')
                         ->with('success','Post created successfully.');
 		}	 
        
     }
+	
+	public function view($slug = null) {
+		
+		$user = Auth::user();
+		$post = Post::where('slug',$slug)->first();				
+        if(Auth::check()){		
+			$user = User::find($user->id);   
+			return view('post_view')->with(compact('user','post'));
+		}else{
+			return view('post_view_without_login')->with(compact('post'));
+		}
+		
+			
+    }	
+	
+	
    
    public function category(Request $request){
 	      $validator = Validator::make($request->all(),[
@@ -79,12 +162,15 @@ class PostController extends Controller {
    }
    
     public function destroy($id)
-    {
-		$post =Post::find($id);
-        $post->delete();
-
-        return redirect()->route('posts')
-                        ->with('success','Post deleted successfully');
+    {	
+		$user_id = Auth::user()->id;
+		$post =Post::find($id);		
+		if($user_id == $post->user_id){			
+			$post->delete();						
+			return redirect()->route('post')->with('success','Post deleted successfully');
+		}else{	
+			return redirect()->route('post')->with('success','Invalid User');		
+		}
     }  
 
 	public function changeStatus(Request $request) {
